Add data_path as an argument and improve error handling
Add a "data_path" argument to several methods in the EikonDataExtractor class to specify the data directory. Improve error handling for when no data is found for specific ISINs, and also update log warnings. A "set_app_key" function is relocated for external access. The version is bumped to 0.0.5, and pytest is added as a development dependency.