Add 'subfolder' parameter to Session class
The 'subfolder' variable is added as a new parameter in the Session class of the eikon_api_wrapper. This addition is used in various parts of the script to change the default subfolder under which the data is stored. Users can now specify a particular subfolder where they want the data to be placed. The default value of this subfolder remains as the original values if not specified.